% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_2_params.R
\docType{data}
\name{example_2_params}
\alias{example_2_params}
\title{Frequency analysis result for 2 parameters distributions, with first column as observations.}
\format{
A data frame with 48 rows and 11 variables:
\describe{
\item{Obs}{observations}
\item{weibull}{weibull}
\item{norm}{normal}
\item{lnorm}{log normal}
\item{gamma}{2 parameters gamma}
\item{gumbel}{gumbel}
\item{logis}{logistic}
\item{cauchy}{cauchy}
\item{llogis}{log logistic}
\item{pareto}{pareto}
\item{trgamma}{transformed gamma}
}
}
\usage{
example_2_params
}
\description{
A dataset containing result of a frequency analysis for 2 parameters distributions.
}
\keyword{datasets}
