% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_3_params.R
\docType{data}
\name{example_3_params}
\alias{example_3_params}
\title{Frequency analysis result for 3 parameters distributions, with first column as observations.}
\format{
A data frame with 48 rows and 5 variables:
\describe{
\item{Obs}{observations}
\item{gamma3}{3 parameters gamma}
\item{gev}{gev}
\item{lgamma3}{3 parameters log gamma}
\item{lnorm3}{3 parameters log normal}
}
}
\usage{
example_3_params
}
\description{
A dataset containing result of a frequency analysis for 3 parameters distributions.
}
\keyword{datasets}
