% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_2_params.R
\docType{data}
\name{fit_2_params}
\alias{fit_2_params}
\title{Distributions fitted with 2 parameters}
\format{
A list with 10 lists of class mle:
\describe{
\item{weibull}{weibull}
\item{norm}{normal}
\item{lnorm}{log normal}
\item{gamma}{2 parameters gamma}
\item{gumbel}{gumbel}
\item{logis}{logistic}
\item{cauchy}{cauchy}
\item{llogis}{log logistic}
\item{pareto}{pareto}
\item{trgamma}{transformed gamma}
}
}
\usage{
fit_2_params
}
\description{
A dataset containing result of fit for 2 parameters distributions.
}
\keyword{datasets}
