% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_3_params.R
\docType{data}
\name{fit_3_params}
\alias{fit_3_params}
\title{Distributions fitted with 3 parameters}
\format{
A list with 4 lists of class mle:
\describe{
\item{gamma3}{3 parameters gamma}
\item{gev}{gev}
\item{lgamma3}{3 parameters log gamma}
\item{lnorm3}{3 parameters log normal}
}
}
\usage{
fit_3_params
}
\description{
A dataset containing result of fit for 3 parameters distributions.
}
\keyword{datasets}
