% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroCHISQR.R
\name{hydroCHISQR}
\alias{hydroCHISQR}
\title{Chi - Squared test for hydrological frequency analysis}
\usage{
hydroCHISQR(df, nc, dist_param, alpha)
}
\arguments{
\item{df}{Data frame with observations and distributions adjusted}

\item{nc}{Number of classes}

\item{dist_param}{Number of parameters of the distributions in df}

\item{alpha}{Significance level}
}
\description{
Chi - Squared test for hydrological frequency analysis
}
